Read audio properties and system file properties
These properties are numeric but it is not worth the hassle to store
them as a numeric type. After all, these values will only be displayed
to the user. It is easier to keep maintain type safety in Kotlin code by
storing these properties as strings in the HashMap and later casting
them as required. Any other option seems to be over engineering.